+++ src/ducks/Signals/chart/SignalPreview.module.scss
@@ -69,6 +69,7 @@
 }
 
 .tooltip {
+  min-width: 320px;
   background: var(--white);
 }
